// src/types.d.ts
export interface Product {
  _id: string;
  name: string;
  description?: string;
  price: number;
  inventory: {
    managed: boolean;
    avaible?: number;
  };
  conditionals: {
    isActive: boolean;
    isInventoryManaged: boolean; // resolverdb
    isSoldOut: boolean; //resolverdb
    hasImages: boolean; //resolverdb
  };
  image?: string;
  assets?: string[];
  sales: number; //resolverdb
  variantGroups?: VariantGroup[]; //resolverdb
  variants?: Variant[]; //resolverdb
  shippingMethods: ShippingMethod[];
  categories?: Category[];
  relatedProducts?: Product[];
  discount?: Discount; //resolverdb
  createdAt: Date;
  updatedAt: Date;
}

export interface createProductInput {
  name: string;
  description?: string;
  price: number;
  inventory: {
    managed: boolean;
    avaible?: number;
  };
  conditionals: {
    isActive: boolean;
  };
  image?: string;
  assets?: string[];
  shippingMethods: string[];
  categories?: string[];
  relatedProducts?: string[];
}

// VariantGroup
export interface VariantGroup {
  _id: string;
  name: string;
  options: string[];
  assets?: string[];
  productId: string;
}

export interface createVariantGroupInput {
  name: string;
  options: string[];
  assets?: string[];
  productId: string;
}

export interface updateVariantGroupInput {
  _id: String;
  name?: string;
  options?: string[];
  assets?: string[];
}

//Variant
export interface Variant {
  _id: string;
  inventory?: number;
  price: number;
  description?: string;
  options: {
    variantGroup: string;
    option: string;
  }[];
  assets?: string[];
  productId: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface createVariantInput {
  inventory?: number;
  price?: number;
  description?: string;
  options: {
    variantGroup: string;
    option: string;
  }[];
  assets?: string[];
  productId: string;
}

export interface updateVariantInput {
  _id: string;
  inventory?: number;
  price?: number;
  description?: string;
  assets?: string[];
}

//Category
export interface Category {
  _id: string;
  parent?: Category;
  name: string;
  description?: string;
  assets?: string[];
  children?: Category[]; // resolver
  products: Product[]; // resolver
  createdAt: Date;
  updatedAt: Date;
}
// ...
export interface ShippingMethod {
  _id: string;
  name: string;
  price: number;
  zones: {
    country: string;
    regions: string[];
  }[];
}

export interface shippingMethodInput {
  name: string;
  price: number;
  zones: {
    country: string;
    regions: string[];
  }[];
}

//Discount
export interface Discount {
  _id: string;
  code: string;
  type: "percentage" | "number";
  value: number;
  startsOn: Date;
  endsOn?: Date;
  isExpired: boolean; //server
  limitQuantity?: number;
  currentQuantity: number; //server
  products: Product[];
}